requirements.txt
dnspython 
pandas 
openpyxl